Changed fastpc to use sort ratio as a double. Updated parse_output to print sort ratio

+++ code_mps/fastpc.h
@@ -60,7 +60,7 @@ private:
   
   string file_name;
   double eps, epsilon;
-  int sort_ratio;  //factor by which elements are pseudo-sorted
+  float sort_ratio;  //factor by which elements are pseudo-sorted
 
   // To implement random_pair function in paper, we must compute |p_pXuh||p_d|/(|p_pXuh||p_d|+|p_p||p_dXu|)
   // Because of normalization, sampler weights at any given time during alg are internally consistent,
@@ -83,7 +83,7 @@ private:
   void freeze_and_sample(my_vector<line_element>& M, my_vector<line_element>& MT, int rows, int cols, dual_sampler_t* p_d, primal_sampler_t* p_p, dual_u_sampler_t* p_dXu, primal_u_sampler_t* p_pXuh, double epsilon, int prob);
 
 public:	
-  solve_instance(double epsilon, string file_name, int sort_ratio);
+  solve_instance(double epsilon, string file_name, float sort_ratio);
   void solve();
 
   void bound_sort(); //sort or pseudo-sort M and MT based on sorting factor
